FastAPIdoc/requirements.txt
mecab-python3==0.996.5
scikit-learn
nltk
fastapi
uvicorn
ipadic
janome
pandas
